+++ packages/core/src/session/runner/llm.ts
@@ -1,21 +1,18 @@
-import { LLM, LLMClient, LLMError, LLMEvent, SystemPart } from "@opencode-ai/llm"
-import { Cause, DateTime, Effect, FiberSet, Layer, Semaphore, Stream } from "effect"
+import { LLM, LLMClient, SystemPart } from "@opencode-ai/llm"
+import { DateTime, Effect, Layer } from "effect"
 import { EventV2 } from "../../event"
-import { ModelV2 } from "../../model"
-import { ProviderV2 } from "../../provider"
 import { SessionSchema } from "../schema"
 import { SessionEvent } from "../event"
 import { SessionStore } from "../store"
 import { Service, StepLimitExceededError } from "./index"
-import { createLLMEventPublisher } from "./publish-llm-event"
 import { toLLMMessages } from "./to-llm-message"
 import { ToolRegistry } from "../../tool/registry"
 import { SessionRunnerModel } from "./model"
 import { Database } from "../../database/database"
 import { SessionInput } from "../input"
-import { QuestionV2 } from "../../question"
 import { SystemContextRegistry } from "../../system-context-registry"
 import { SessionContextEpoch } from "../context-epoch"
+import { SessionRunnerProviderTurn } from "./provider-turn"
 
 /**
  * Runs one durable coding-agent Session until it settles.
@@ -68,7 +65,7 @@ import { SessionContextEpoch } from "../context-epoch"
  *   - [ ] Coalesce streamed deltas and add covering projected-history indexes.
  *   - [ ] Update title, summaries, compaction state, and cleanup in bounded background work.
  *
- * Use `llm.stream(request)` for each provider turn. Keep tool execution and continuation here.
+ * Use `llm.stream(request)` for each provider turn. Keep activity continuation here.
  * Durable activity recovery remains a separate future slice with an explicit retry policy.
  *
  * The current slice loads V2 history, translates it, resolves a model through a core service, and persists one
@@ -89,19 +86,17 @@ export const layer = Layer.effect(
     const store = yield* SessionStore.Service
     const systemContext = yield* SystemContextRegistry.Service
     const db = (yield* Database.Service).db
+    const runProviderTurn = SessionRunnerProviderTurn.make({ events, llm, tools })
     const getSession = Effect.fn("SessionRunner.getSession")(function* (sessionID: SessionSchema.ID) {
       const session = yield* store.get(sessionID)
       if (!session) return yield* Effect.die(`Session not found: ${sessionID}`)
       return session
     })
 
-    const getContext = Effect.fn("SessionRunner.getContext")(function* (sessionID: SessionSchema.ID) {
-      return yield* store.context(sessionID)
-    })
     const failInterruptedTools = Effect.fn("SessionRunner.failInterruptedTools")(function* (
       sessionID: SessionSchema.ID,
     ) {
-      for (const message of yield* getContext(sessionID)) {
+      for (const message of yield* store.context(sessionID)) {
         if (message.type !== "assistant") continue
         for (const tool of message.content) {
           if (tool.type !== "tool" || (tool.state.status !== "pending" && tool.state.status !== "running")) continue
@@ -120,22 +115,13 @@ export const layer = Layer.effect(
       }
     })
 
-    const awaitToolFibers = (fibers: FiberSet.FiberSet<void, never>) =>
-      Effect.raceFirst(FiberSet.join(fibers), FiberSet.awaitEmpty(fibers))
-
-    // Match V1: dismissing a question halts the loop instead of becoming model-facing tool output.
-    const isQuestionRejected = (cause: Cause.Cause<unknown>) =>
-      cause.reasons.some((reason) => Cause.isDieReason(reason) && reason.defect instanceof QuestionV2.RejectedError)
-
-    const runTurn = Effect.fn("SessionRunner.runTurn")(function* (
+    const prepareProviderTurn = Effect.fn("SessionRunner.prepareProviderTurn")(function* (
       sessionID: SessionSchema.ID,
-      promotion: "steer" | "queue" | undefined,
+      promotion: SessionInput.Delivery | undefined,
     ) {
       const session = yield* getSession(sessionID)
       const initialized = yield* SessionContextEpoch.initialize(db, systemContext, session.id, session.location)
       const model = yield* models.resolve(session)
-      const toolFibers = yield* FiberSet.make<void, never>()
-      let needsContinuation = false
       if (promotion) {
         const cutoff = yield* SessionInput.latestSeq(db, session.id)
         if (promotion === "steer") yield* SessionInput.promoteSteers(db, events, session.id, cutoff)
@@ -146,99 +132,32 @@ export const layer = Layer.effect(
       }
       const system = initialized ?? (yield* SessionContextEpoch.prepare(db, events, systemContext, session.id, session.location))
       const context = yield* store.runnerContext(session.id, system.baselineSeq)
-      const request = LLM.request({
-        model,
-        system: system.baseline.length > 0 ? [SystemPart.make(system.baseline)] : [],
-        messages: toLLMMessages(context, model),
-        tools: yield* tools.definitions(),
-      })
-      const publisher = createLLMEventPublisher(events, {
+      return {
         sessionID: session.id,
         agent: session.agent ?? "build",
-        model: {
-          id: ModelV2.ID.make(model.id),
-          providerID: ProviderV2.ID.make(model.provider),
-          ...(session.model?.variant === undefined ? {} : { variant: session.model.variant }),
-        },
-      })
-      const withPublication = Semaphore.makeUnsafe(1).withPermit
-      const publish = (event: LLMEvent) => withPublication(publisher.publish(event))
-      const providerStream = llm.stream(request).pipe(
-        Stream.runForEach((event) =>
-          Effect.gen(function* () {
-            yield* publish(event)
-            if (event.type !== "tool-call" || event.providerExecuted) return
-            needsContinuation = true
-            yield* tools.settle({ sessionID: session.id, call: event }).pipe(
-              Effect.catchCause((cause) => {
-                if (isQuestionRejected(cause)) return Effect.failCause(cause)
-                return Effect.succeed({
-                  result: { type: "error" as const, value: String(Cause.squash(cause)) },
-                  output: undefined,
-                })
-              }),
-              Effect.flatMap((settlement) =>
-                publish(
-                  LLMEvent.toolResult({
-                    id: event.id,
-                    name: event.name,
-                    result: settlement.result,
-                    output: settlement.output,
-                  }),
-                ),
-              ),
-              FiberSet.run(toolFibers),
-            )
-          }),
-        ),
-        Effect.ensuring(withPublication(publisher.flush())),
-      )
-
-      return yield* Effect.uninterruptibleMask((restore) =>
-        Effect.gen(function* () {
-          const stream = yield* restore(providerStream).pipe(Effect.exit)
-          let llmFailure: LLMError | undefined
-          if (stream._tag === "Failure") {
-            for (const reason of stream.cause.reasons) {
-              if (!Cause.isFailReason(reason)) continue
-              if (reason.error instanceof LLMError) llmFailure = reason.error
-            }
-          }
-          if (llmFailure && !publisher.hasProviderError()) {
-            yield* withPublication(publisher.failUnsettledTools("Provider did not return a tool result", true))
-            yield* withPublication(
-              events.publish(SessionEvent.Step.Failed, {
-                sessionID: session.id,
-                timestamp: yield* DateTime.now,
-                assistantMessageID: yield* publisher.startAssistant(),
-                error: { type: "unknown", message: llmFailure.reason.message },
-              }),
-            )
-          }
-          if (stream._tag === "Failure" && Cause.hasInterrupts(stream.cause)) yield* FiberSet.clear(toolFibers)
-          const settled = yield* restore(awaitToolFibers(toolFibers)).pipe(Effect.exit)
-          if (settled._tag === "Failure" && isQuestionRejected(settled.cause)) {
-            yield* FiberSet.clear(toolFibers)
-            yield* withPublication(publisher.failUnsettledTools("Tool execution interrupted"))
-            return yield* Effect.interrupt
-          }
-          if (
-            (stream._tag === "Failure" && Cause.hasInterrupts(stream.cause)) ||
-            (settled._tag === "Failure" && Cause.hasInterrupts(settled.cause))
-          ) {
-            yield* FiberSet.clear(toolFibers)
-            yield* withPublication(publisher.failUnsettledTools("Tool execution interrupted"))
-          }
-          if (publisher.hasProviderError())
-            yield* withPublication(publisher.failUnsettledTools("Tool execution interrupted"))
-          if (stream._tag === "Success" && !publisher.hasProviderError())
-            yield* withPublication(publisher.failUnsettledTools("Provider did not return a tool result", true))
-          const attempt = stream._tag === "Failure" ? stream : settled
-          if (attempt._tag === "Failure") return yield* Effect.failCause(attempt.cause)
-          return !publisher.hasProviderError() && needsContinuation
+        ...(session.model?.variant === undefined ? {} : { variant: session.model.variant }),
+        request: LLM.request({
+          model,
+          system: system.baseline.length > 0 ? [SystemPart.make(system.baseline)] : [],
+          messages: toLLMMessages(context, model),
+          tools: yield* tools.definitions(),
         }),
-      )
-    }, Effect.scoped)
+      }
+    })
+
+    const runActivity = Effect.fn("SessionRunner.runActivity")(function* (
+      sessionID: SessionSchema.ID,
+      initialPromotion: SessionInput.Delivery | undefined,
+    ) {
+      let promotion = initialPromotion
+      for (let step = 0; step < MAX_STEPS; step++) {
+        const needsContinuation = yield* prepareProviderTurn(sessionID, promotion).pipe(Effect.flatMap(runProviderTurn))
+        promotion = "steer"
+        if (needsContinuation || (yield* SessionInput.hasPending(db, sessionID, "steer"))) continue
+        return
+      }
+      return yield* new StepLimitExceededError({ sessionID, limit: MAX_STEPS })
+    })
 
     const run = Effect.fn("SessionRunner.run")(function* (input: {
       readonly sessionID: SessionSchema.ID
@@ -248,21 +167,8 @@ export const layer = Layer.effect(
       const hasQueue = hasSteer ? false : yield* SessionInput.hasPending(db, input.sessionID, "queue")
       if (input.force !== true && !hasSteer && !hasQueue) return
       yield* failInterruptedTools(input.sessionID)
-      let promotion: "steer" | "queue" | undefined = hasSteer ? "steer" : hasQueue ? "queue" : undefined
-      let openActivity = input.force === true || hasSteer || hasQueue
-      while (openActivity) {
-        let needsContinuation = true
-        for (let step = 0; step < MAX_STEPS; step++) {
-          needsContinuation = yield* runTurn(input.sessionID, promotion)
-          promotion = "steer"
-          if (!needsContinuation) needsContinuation = yield* SessionInput.hasPending(db, input.sessionID, "steer")
-          if (!needsContinuation) break
-        }
-        if (needsContinuation)
-          return yield* new StepLimitExceededError({ sessionID: input.sessionID, limit: MAX_STEPS })
-        openActivity = yield* SessionInput.hasPending(db, input.sessionID, "queue")
-        promotion = openActivity ? "queue" : undefined
-      }
+      yield* runActivity(input.sessionID, hasSteer ? "steer" : hasQueue ? "queue" : undefined)
+      while (yield* SessionInput.hasPending(db, input.sessionID, "queue")) yield* runActivity(input.sessionID, "queue")
     })
 
     return Service.of({
